Broke down updating DailyEmissions
